fix(#162): drop setMigrationHandler to keep Grafana API compatibility gate green
The levitate CI gate flags PanelPlugin.setMigrationHandler because its
signature changed between @grafana/data 11.x and 13.x. The handler was
belt-and-braces: the WeathermapBuilder render fix alone resolves the
editor crash, and WeathermapPanel already migrates during panel render.

Move the migration coverage to direct handleVersionedStateUpdates unit
tests in utils.test.ts and share the legacy pre-v14 fixture via testData.
